@@ -1,43 +1,42 @@
 using Blazored.Toast.Services;
 using Xunit;
 
-namespace Blazored.Toast.Tests.ToastServiceTests
+namespace Blazored.Toast.Tests.ToastServiceTests;
+
+public class ClearInfoToasts
 {
-    public class ClearInfoToasts
+    private readonly ToastService _sut;
+
+    public ClearInfoToasts()
     {
-        private readonly ToastService _sut;
+        _sut = new ToastService();
+    }
 
-        public ClearInfoToasts()
-        {
-            _sut = new ToastService();
-        }
+    [Fact]
+    public void OnClearToastsInnvoked_When_ClearInfoToastsCalled()
+    {
+        // arrange
+        var OnClearToastsCalled = false;
+        _sut.OnClearToasts += (_) => OnClearToastsCalled = true;
 
-        [Fact]
-        public void OnClearToastsInnvoked_When_ClearInfoToastsCalled()
-        {
-            // arrange
-            var OnClearToastsCalled = false;
-            _sut.OnClearToasts += (_) => OnClearToastsCalled = true;
+        // act
+        _sut.ClearInfoToasts();
 
-            // act
-            _sut.ClearInfoToasts();
+        // assert
+        Assert.True(OnClearToastsCalled);
+    }
 
-            // assert
-            Assert.True(OnClearToastsCalled);
-        }
+    [Fact]
+    public void OnClearToastsContainsToastLevelWarning_When_ClearInfoToastsCalled()
+    {
+        // arrange
+        var toastLevel = "";
+        _sut.OnClearToasts += (argToastlevel) => toastLevel = argToastlevel.ToString();
 
-        [Fact]
-        public void OnClearToastsContainsToastLevelWarning_When_ClearInfoToastsCalled()
-        {
-            // arrange
-            var toastLevel = "";
-            _sut.OnClearToasts += (argToastlevel) => toastLevel = argToastlevel.ToString();
+        // act
+        _sut.ClearInfoToasts();
 
-            // act
-            _sut.ClearInfoToasts();
-
-            // assert
-            Assert.Equal(ToastLevel.Info.ToString(), toastLevel);
-        }
+        // assert
+        Assert.Equal(ToastLevel.Info.ToString(), toastLevel);
     }
 }
